requirements.txt
annotated-types==0.6.0
Faker==23.2.1
greenlet==3.0.3
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2
psycopg==3.1.17
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
